# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EMS API Documentation",
        "description": "Dokumentasi API untuk event management system:<br>Auth: Register, Login, dan Logout<br>Event: daftar event, detail event, pendaftaran user<br>Attendance: check-in, check-out, dan riwayat",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server utama API EMS"
        }
    ],
    "paths": {
...
        "/api/v1/events/{id}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Lihat detail event",
                "description": "Mengambil detail lengkap dari event berdasarkan ID.",
                "operationId": "1b3968b8db00da68b19911a6edf1ba81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID event yang ingin dilihat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail event ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Laravel Developer Conference 2025"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Event tahunan untuk para developer Laravel."
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Jakarta Convention Center"
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-11-01"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-11-03"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event tidak ditemukan"
                    }
                }
            }
        },
        "/api/v1/events/{id}/register": {
            "post": {
                "tags": [
                    "Event"
                ],
                "summary": "Daftar ke event",
                "description": "User yang sedang login melakukan pendaftaran ke event berdasarkan ID.",
                "operationId": "f8ed3ae3bed686bdee8fbbb6f0589a2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID event yang ingin didaftarkan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pendaftaran event berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registered"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "event_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "registration_time": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2025-10-16T09:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Event tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/events/user-registered": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Daftar event yang diikuti user",
                "description": "Mengambil semua event yang sudah didaftarkan oleh user yang sedang login.",
                "operationId": "5d8357f78f38fc06274ca5fa51d8f39e",
                "responses": {
                    "200": {
                        "description": "Daftar event user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "event_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "registration_time": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2025-10-16T09:30:00Z"
                                            },
                                            "event": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Laravel Developer Conference 2025"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Jakarta Convention Center"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}